Do not override window.ResizeObserver
The polyfill does not match the spec https://drafts.csswg.org/resize-observer-1/

This is the implementation used by `d2l-card`, which `d2l-table` is overriding
https://www.npmjs.com/package/resize-observer-polyfill

Since `d2l-tspan-resize-observer` observer

**Steps to Reproduce**

1. Open Edge or IE11
1. Add an observer
1. Receive callback

**Expected**

- `arguments[0]` is an array of `entries` (as defined in the specificiation)

**Actual**

- `arguments[0]` is an `HTMLElement`
