+++ d2l-tspan-resize-observer.js
@@ -1,0 +1,43 @@
+var ResizeObserver = function(callback) {
+	this._callback = callback;
+	this._cancellationTokens = [];
+};
+
+ResizeObserver.prototype.observe = function(node) {
+	var cancellationToken = {
+		ref: node,
+		cancelled: false
+	};
+	this._cancellationTokens.push(cancellationToken);
+	this._callback(node);
+
+	var lastWidth = node.offsetWidth;
+	var lastHeight = node.offsetHeight;
+	var _checkForChanges = function() {
+		if (
+			node.offsetWidth !== lastWidth ||
+			node.offsetHeight !== lastHeight
+		) {
+			lastWidth = node.offsetWidth;
+			lastHeight = node.offsetHeight;
+			this._callback(node);
+		}
+
+		if (!cancellationToken.cancelled) {
+			window.requestAnimationFrame(_checkForChanges);
+		}
+	}.bind(this);
+	_checkForChanges();
+};
+
+ResizeObserver.prototype.unobserve = function(node) {
+	for (var i = 0; i < this._cancellationTokens.length; i++) {
+		if (this._cancellationTokens[i].ref === node) {
+			this._cancellationTokens[i].cancelled = true;
+			this._cancellationTokens.splice(i, 1);
+			break;
+		}
+	}
+};
+
+export default ResizeObserver;
